feat: added piepline codebuild updater handler func tests
